// app.client/src/api/models/DistributorListVM.ts
/* generated manually */

export type DistributorListVM = {
    distributorId?: string | null;
    name?: string | null;
    email?: string | null;
    phone?: string | null;
    website?: string | null;
    gstNumber?: string | null;
    panNumber?: string | null;
    address?: string | null;
    note?: string | null;
    isVoided?: boolean;
    createdBy?: string | null;
    createdDate?: string;
    lastModifiedBy?: string | null;
    lastModifiedDate?: string | null;
};